refactor: cleanup packages/constants
